% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHandle.R
\name{GetCurrWrd}
\alias{GetCurrWrd}
\title{Get a Handle to a Running Word/Excel/PowerPoint Instance}
\usage{
GetCurrWrd()
}
\value{
a handle (pointer) to the running Word, resp. Excel instance.
}
\description{
Look for a running Word, resp. Excel instance and return its handle. If no
running instance is found a new instance will be created (which will be
communicated with a warning).
}
\note{
When closing an application instance, the value of the pointer in R is
not somehow automatically invalidated. In such cases the corresponding
variable contains an invalid address.  Whether the pointer still refers to a
valid running application instance can be checked by
\code{\link{IsValidHwnd}}.
}
\examples{

\dontrun{# Windows-specific example

# Start a new instance
GetNewWrd()

# grab the handle to this instance
wrd <- GetCurrWrd()

# this should be valid
IsValidHwnd(wrd)

# close the instance
wrd$quit()

# now it should be gone and the pointer invalid
if(IsValidHwnd(wrd)){ 
  print("Ouups! Still there?")
} else {  
  print("GetCurrWrd: no running word instance found...")
}
}

}
\seealso{
\code{\link{GetNewWrd}}, \code{\link{IsValidHwnd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{misc}
