% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHandle.R
\name{GetNewWrd}
\alias{GetNewWrd}
\alias{GetNewXL}
\alias{GetNewPP}
\alias{GetCurrWrd}
\alias{GetCurrXL}
\alias{GetCurrPP}
\alias{GetHwnd}
\title{Get a Handle to a Word/Excel/PowerPoint Instance}
\usage{
GetNewWrd(visible = TRUE, newdoc = TRUE, template = "Normal")

GetNewXL(visible = TRUE, newdoc = TRUE)

GetNewPP(visible = TRUE)

GetCurrWrd()

GetCurrXL()

GetCurrPP()
}
\arguments{
\item{visible}{logical, should the application made visible? Defaults to \code{TRUE}.}

\item{newdoc}{logical, determining if a new document
should be created. Defaults to \code{TRUE}.}

\item{template}{the name of the template to be used for creating a new
Word document. Ignored by Excel and PowerPoint.}
}
\value{
a handle (pointer) to the running Word/Excel/PowerPoint instance.
}
\description{
To be able to control MS-Office applications we need a so called handle,
a kind of address, to it. It is either possible to get the handle of an
already running application or to start a new one.
}
\details{
The functions \code{GetNewWrd()}, \code{GetNewXL()}, and \code{GetNewPP()}
launch a new instance of the required application and return its handle.

The functions \code{GetCurrWrd()}, \code{GetCurrXL()}, and \code{GetCurrPP()}
look for a running instance of the specific application and return its handle.
Unfortunately it is not possible to choose a specific handle if there are
several instances already running. The underlying RDCOM function yields
some kind of "latest" launched instance. So in most cases you will be
safer to start a new instance to be sure that you're communicating
with the expected instance. If there's no running instance the error
message will be returned with a warning.
}
\note{
When closing an application instance, the value of the pointer in R is
not somehow automatically invalidated. In such cases the corresponding
variable contains an invalid address.  Whether the pointer still refers to a
valid running application instance can be checked by
\code{\link{IsValidHwnd}}.
}
\examples{

\dontrun{# Windows-specific example

# Start a new instance
wrd <- GetNewWrd()
WrdText("Send some Text", wrd=wrd)

# Release the handle
rm(wrd)

# regrab the handle to this instance
wrd <- GetCurrWrd()

# this should be valid
IsValidHwnd(wrd)

# close the instance
wrd$quit()

# now it should be gone and the pointer invalid
if(IsValidHwnd(wrd)){ 
  print("Ouups! Still there?")
} else {  
  print("GetCurrWrd: no running word instance found...")
}


# Same with PowerPoint:
# get a handle to a new PowerPoint instance
pp <- GetNewPP()

# send some text
PpText("Some text on a slide!\n", 
       fontname="Arial", x=200, y=70, height=30, fontsize=14,
       bold=TRUE, pp=pp, bg="lemonchiffon", hasFrame=TRUE)

}
}
\seealso{
\code{\link{GetNewWrd}}, \code{\link{IsValidHwnd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{misc}
