% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHandle.R
\name{GetNewXL}
\alias{GetNewXL}
\alias{GetCurrXL}
\alias{GetNewPP}
\alias{GetCurrPP}
\alias{XLKill}
\title{Create a New Excel Instance}
\usage{
GetNewXL(visible = TRUE, newdoc = TRUE)

GetCurrXL()

GetNewPP(visible = TRUE)

GetCurrPP()

XLKill()
}
\arguments{
\item{visible}{logical, should Excel made visible? Defaults to \code{TRUE}.}

\item{newdoc}{logical, determining if a new workbook should be created.
Defaults to \code{TRUE}.}
}
\description{
Start a new instance of Excel and return its handle. This is needed to
address the Excel application and objects afterwards.
}
\details{
Here's a list of some frequently used commands.\cr Let's assume:
\preformatted{xl <- GetNewXL() } \tabular{ll}{ workbooks \tab
\code{xl$workbooks()$count()} \cr quit excel \tab \code{xl$quit()} \cr }

#' \code{XLKill} will kill a running XL instance (which might be invisible).
Background is the fact, that the simple XL$quit() command would not
terminate a running XL task, but only set it invisible (observe the
TaskManager). This ghost version may sometimes confuse XLView and hinder to
create a new instance. In such cases you have to do the garbage
collection...
}
\examples{

\dontrun{ # Windows-specific example
# get a handle to a new excel instance
xl <- GetNewXL()
}

}
\seealso{
\code{\link{XLView}}, \code{\link{XLGetRange}},
\code{\link{XLGetWorkbook}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{misc}
