% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHandle.R
\name{IsValidHwnd}
\alias{IsValidHwnd}
\title{Check Windows Pointer}
\usage{
IsValidHwnd(hwnd)
}
\arguments{
\item{hwnd}{the pointer to an application instance as created by
\code{GetCOMAppHandle()}.}
}
\value{
logical value, \code{TRUE} if \code{hwnd} is a valid pointer to a running
application
}
\description{
Check if a pointer points to a valid and running MS-Office instance. The
function does this by first checking for \code{NULL} pointer
and then trying to get the current selection of the application.
}
\note{
When closing an application instance, the value of the pointer in R is
not somehow automatically invalidated. In such cases the corresponding
variable contains an invalid address.  Whether the pointer still refers to a
valid running application instance can be checked by
\code{\link{IsValidHwnd}}.
}
\seealso{
\code{\link{GetHwnd}()}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
