% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlook.R
\name{SendOutlookMail}
\alias{SendOutlookMail}
\title{Send a Mail Using Outlook as Mail Client}
\usage{
SendOutlookMail(to, cc = NULL, bcc = NULL, subject, body, attachment = NULL)
}
\arguments{
\item{to}{a vector of recipients}

\item{cc}{a vector of recipients receiving a carbon copy}

\item{bcc}{a vector of recipients receiving a blind carbon copy}

\item{subject}{the subject of the mail}

\item{body}{the body text of the mail}

\item{attachment}{a vector of paths to attachments}
}
\value{
\code{TRUE} is returned in case of success and \code{FALSE} in case of an
unhandled error while sending mail
}
\description{
Sending emails in R can be required in some reporting tasks. As we already
have RDCOMClient available we wrap the send code in a function.
}
\examples{
\dontrun{
SendOutlookMail(to=c("me@microsoft.com", "you@rstudio.com"), subject = "Some Info", 
                body = "Hi all\r Find the files attached\r Regards, Andri", 
                attachment = c("C:/temp/fileA.txt", 
                               "C:/temp/fileB.txt"))
}

}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net} strongly based on code of
Franziska Mueller
}
\keyword{MS-Office}
