% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdCellRange}
\alias{WrdCellRange}
\title{Return the Cell Range Of a Word Table}
\usage{
WrdCellRange(wtab, from, to)
}
\arguments{
\item{wtab}{a handle to the word table as returned i.g. by
\code{\link{WrdTable}}}

\item{from}{a vector containing row- and column number of the left/upper
cell of the cell range.}

\item{to}{a vector containing row- and column number of the right/lower cell
of the cell range.}
}
\value{
a handle to the range.
}
\description{
Return a handle of a cell range of a word table. This is useful for
formating the cell range.
}
\details{
Cell range selecting might be complicated. This function makes it easy.
}
\examples{

\dontrun{

# Windows-specific example
wrd <- GetNewWrd()
WrdTable(nrow=3, ncol=3, wrd=wrd)
crng <- WrdCellRange(from=c(1,2), to=c(2,3))
crng$Select()
}

}
\seealso{
\code{\link{WrdTable}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
