% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdFormatCells}
\alias{WrdFormatCells}
\title{Format Cells Of a Word Table}
\usage{
WrdFormatCells(
  wtab,
  rstart,
  rend,
  col = NULL,
  bg = NULL,
  font = NULL,
  border = NULL,
  align = NULL
)
}
\arguments{
\item{wtab}{a handle to the word table as returned i.g. by
\code{\link{WrdTable}}}

\item{rstart}{the left/upper cell of the cell range}

\item{rend}{the right/lower cell of the cell range}

\item{col}{the foreground colour}

\item{bg}{the background colour}

\item{font}{the font to be used to the output. This should be defined as a
list containing fontname, fontsize, bold and italic flags:\cr
\code{list(name="Arial", size=10, bold=FALSE, italic=TRUE,
color=wdConst$wdColorBlack)}.}

\item{border}{the border of the cell range, defined as a list containing
arguments for border, linestyle, linewidth and color. \code{border} is a
vector containing the parts of the border defined by the Word constants
\code{wdConst$wdBorder...}, being $wdBorderBottom, $wdBorderLeft,
$wdBorderTop, $wdBorderRight, $wdBorderHorizontal, $wdBorderVertical,
$wdBorderDiagonalUp, $wdBorderDiagonalDown. linestyle, linewidth and color
will be recycled to the required dimension.}

\item{align}{a character out of \code{"l"}, \code{"c"}, \code{"r"} setting
the horizontal alignment of the cell range.}
}
\value{
a handle to the range.
}
\description{
Format cells of a Word table.
}
\details{
Cell range selecting might be complicated. This function makes it easy.
}
\examples{

\dontrun{   # Windows-specific example

m <- matrix(rnorm(12)*100, nrow=4,
            dimnames=list(LETTERS[1:4], c("Variable","Value","Remark")))

wrd <- GetNewWrd()
wt <- ToWrd(m)

WrdFormatCells(wt, rstart=c(3,1), rend=c(4,3),
               bg=wdConst$wdColorGold, font=list(name="Arial Narrow", bold=TRUE),
               align="c", border=list(color=wdConst$wdColorTeal,
                                      linewidth=wdConst$wdLineWidth300pt))

}

}
\seealso{
\code{\link{WrdTable}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
