% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrdBookmark.R
\name{WrdInsertBookmark}
\alias{WrdInsertBookmark}
\alias{WrdDeleteBookmark}
\title{Insert New/Delete MS-Word Bookmark
\code{WrdInsertBookmark}, \code{WrdDeleteBookmark} inserts/deletes
a bookmark in a Word document.}
\usage{
WrdInsertBookmark(name, wrd = DescToolsOfficeOptions("lastWord"))

WrdDeleteBookmark(name, wrd = DescToolsOfficeOptions("lastWord"))
}
\arguments{
\item{name}{the name of the bookmark.}

\item{wrd}{the pointer to a word instance. Can be a new one, created by
\code{GetNewWrd()} or an existing one, created by \code{GetCurrWrd()}.
Default is the last created pointer stored in
\code{DescToolsOfficeOptions("lastWord")}.}
}
\description{
Insert New/Delete MS-Word Bookmark
\code{WrdInsertBookmark}, \code{WrdDeleteBookmark} inserts/deletes
a bookmark in a Word document.
}
\examples{

\dontrun{ # we can't get this through the CRAN test - run it with copy/paste to console
wrd <- GetNewWrd()
WrdText("a)\n\n\nb)", fontname=WrdGetFont()$name, fontsize=WrdGetFont()$size)
WrdInsertBookmark("chap_b")
WrdText("\n\n\nc)\n\n\n", fontname=WrdGetFont()$name, fontsize=WrdGetFont()$size)

WrdGoto("chap_b")
WrdUpdateBookmark("chap_b", "Goto chapter B and set text")

WrdInsertBookmark("mybookmark")
ToWrd("A longer text\n\n\n")

# Now returning the bookmark
bm <- WrdBookmark("mybookmark")

# get the automatically created name of the bookmark
bm$name()
}
}
\seealso{
\code{\link{WrdFont}}, \code{\link{WrdPlot}},
\code{\link{GetNewWrd}}, \code{\link{GetCurrWrd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
