% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdPageBreak}
\alias{WrdPageBreak}
\title{Insert a Page Break}
\usage{
WrdPageBreak(wrd = DescToolsOptions("lastWord"))
}
\arguments{
\item{wrd}{the pointer to a word instance. Can be a new one, created by
\code{GetNewWrd()} or an existing one, created by \code{GetCurrWrd()}.
Default is the last created pointer stored in
\code{DescToolsOptions("lastWord")}.}
}
\description{
Insert a page break in a MS-Word (R) document at the position of the cursor.
}
\examples{

\dontrun{ # Windows-specific example
wrd <- GetNewWrd()
WrdText("This is text on page 1.\n\n")
WrdPageBreak()
WrdText("This is text on another page.\n\n")
}
}
\seealso{
\code{\link{WrdFont}}, \code{\link{WrdPlot}},
\code{\link{GetNewWrd}}, \code{\link{GetCurrWrd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
