% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdSaveAs}
\alias{WrdSaveAs}
\alias{WrdOpenFile}
\title{Open and Save Word Documents}
\usage{
WrdSaveAs(fn, fileformat = "docx", wrd = DescToolsOfficeOptions("lastWord"))
}
\arguments{
\item{fn}{filename and -path for the document.}

\item{fileformat}{file format, one out of \code{"doc"}, \code{"htm"},
\code{"pdf"}.}

\item{wrd}{the pointer to a word instance. Can be a new one, created by
\code{GetNewWrd()} or an existing one, created by \code{GetCurrWrd()}.
Default is the last created pointer stored in
\code{DescToolsOfficeOptions("lastWord")}.}
}
\value{
nothing returned
}
\description{
Open and save MS-Word documents.
}
\examples{

\dontrun{
#   Windows-specific example
wrd <- GetNewWrd()
WrdCaption("A Report")
WrdSaveAs(fn="report", fileformat="htm")
}
}
\seealso{
\code{\link{GetNewWrd}()}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
