% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdStyle}
\alias{WrdStyle}
\alias{WrdStyle<-}
\title{Get or Set the Style in Word}
\usage{
WrdStyle(wrd = DescToolsOptions("lastWord"))

WrdStyle(wrd) <- value
}
\arguments{
\item{wrd}{the pointer to a word instance. Can be a new one, created by
\code{GetNewWrd()} or an existing one, created by \code{GetCurrWrd()}.
Default is the last created pointer stored in
\code{DescToolsOptions("lastWord")}.}

\item{value}{the name of the style to be used to the output. This should be
defined an existing name.}
}
\value{
character, name of the style
}
\description{
\code{WrdStyle} can be used to get and set the style in Word for the text to
be inserted. \code{WrdStyle} returns the style at the current cursor
position.
}
\examples{

\dontrun{ # Windows-specific example

wrd <- GetNewWrd()
# the current stlye
WrdStyle(wrd)
}
}
\seealso{
\code{\link{ToWrd}}, \code{\link{WrdPlot}},
\code{\link{GetNewWrd}}, \code{\link{GetCurrWrd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
