% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{XLSaveAs}
\alias{XLSaveAs}
\title{Save Excel File}
\usage{
XLSaveAs(
  fn,
  file_format = xlConst$XlFileFormat$xlWorkbookNormal,
  xl = DescToolsOptions("lastXL")
)
}
\arguments{
\item{fn}{the filename}

\item{file_format}{the file format using the xl constant.}

\item{xl}{the pointer to a MS-Excel instance. An new instance can be created
with \code{GetNewXL()}, returning the appropriate handle. A handle to an
already running instance is returned by \code{GetCurrXL()}.  Default is the
last created pointer stored in \code{DescToolsOptions("lastXL")}.}
}
\value{
returns \code{TRUE} if the save operation has been successful
}
\description{
Save the current workbook under the given name and format.
}
\examples{

\dontrun{# Windows-specific example
XLView(d.diamonds)
XLSaveAs("Diamonds")
xl$quit()
}
}
\seealso{
\code{\link{XLView}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{manip}
