% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHandle.R
\name{createCOMReference}
\alias{createCOMReference}
\alias{GetCOMAppHandle}
\title{Get a Handle to COM Application}
\usage{
createCOMReference(ref, className)

GetCOMAppHandle(app, option = NULL, existing = FALSE, visible = TRUE)
}
\arguments{
\item{ref}{the S object that is an external pointer containing the reference to the COM object.}

\item{className}{The name of the class that is “suggested” by the caller.}

\item{app}{name of the application as required by COM. This can e.g. be "Word.Application"
for MS-Word.}

\item{option}{logical, should the handle be stored as an option?
If this is left to \code{NULL} no option will be stored, if a text is provided, the
handle will be stored under this name.}

\item{existing}{logical, should the handle to an existing instance be returned?
Defaults to \code{FALSE}.}

\item{visible}{logical, should the application made visible? Defaults to \code{TRUE}.}
}
\value{
a handle (pointer) to a running Word/Excel/PowerPoint or another
COM instance.
}
\description{
To be able to control COM applications we need a so called handle,
a kind of address, to it.
}
\details{
The function \code{GetCOMAppHandle()} is the workhorse for
launching a new instance of the required application and return its handle.
It checks if the RDCOMClient package is available and tries to start the
application. If this is successful the handle is stored as option
}
\examples{

\dontrun{# Windows-specific example

# Start a new instance of an application
hwnd <- GetCOMAppHandle("Word.Application", option = "lastWord", 
                         existing = FALSE, visible = TRUE)

}
}
\seealso{
\code{\link{GetHwnd}}, \code{\link{IsValidHwnd}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{misc}
