% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrd.R
\name{WrdMergeCells}
\alias{WrdMergeCells}
\title{Merges Cells Of a Defined Word Table Range}
\usage{
WrdMergeCells(wtab, rstart, rend)
}
\arguments{
\item{wtab}{a handle to the word table as returned i.g. by
\code{\link{WrdTable}}}

\item{rstart}{the left/upper cell of the cell range.}

\item{rend}{the right/lower cell of the cell range.}
}
\value{
nothing
}
\description{
Merges a cell range of a word table.
}
\examples{

\dontrun{

# Windows-specific example
wrd <- GetNewWrd()
wtab <- WrdTable(nrow=3, ncol=3, wrd=wrd)
WrdMergeCells(wtab, rstart=c(1,2), rend=c(2,3))
}

}
\seealso{
\code{\link{WrdTable}}, \code{\link{WrdCellRange}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{print}
